refactor(netcdf): extract crop cluster into engines/selection.py (STR-1)
Create the Selection(_Engine) collaborator and fold the NetCDF crop
family into it (issue #615): crop and its private helpers
_resolve_crop_mask / _crop_one / _crop_curvilinear move out of netcdf.py.
NetCDF.crop becomes a thin façade delegating to self.selection.crop;
signature, behaviour, and return type are unchanged.

The rectilinear path that the override reached via super().crop(...) is
expressed as self._ds.spatial.crop(...) — the same base Dataset affine
crop, bypassing the NetCDF override. Shared helpers
(_apply_to_all_variables / _preserve_netcdf_metadata / _bbox_geotransform)
stay on NetCDF because the not-yet-moved sel/subset/reduce still use
them; the engine reaches them through self._ds.

Wired as self.selection (no collision with existing attributes).
