+++ src/pyramids/netcdf/engines/selection.py
@@ -1,0 +1,329 @@
+"""Spatial / dimensional selection engine for :class:`pyramids.netcdf.NetCDF`.
+
+Owns the crop / sel / subset / reduce family extracted from the
+``netcdf.py`` god-object (issue #615, STR-1). Per the agreed design the
+NetCDF-specific cropping (``crop`` and its curvilinear path) is folded in
+here rather than living in a separate spatial engine.
+
+The public ``NetCDF`` methods are thin façades delegating to this engine;
+signatures, behaviour, and return types are unchanged. Cropping reaches
+the container's own plumbing through the weakref-proxied back-reference
+``self._ds`` — including ``self._ds.spatial.crop`` for the base affine
+crop (equivalent to the ``super().crop`` call the override used) and the
+shared helpers ``_apply_to_all_variables`` / ``_preserve_netcdf_metadata``
+/ ``_bbox_geotransform`` which stay on ``NetCDF`` because the not-yet-moved
+methods still use them.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+import numpy as np
+from shapely import contains_xy
+
+from pyramids.dataset import DEFAULT_NO_DATA_VALUE
+from pyramids.dataset.engines._base import _Engine
+from pyramids.feature import FeatureCollection
+from pyramids.netcdf._plot import NetCDFPlot
+
+if TYPE_CHECKING:
+    from pyramids.netcdf.netcdf import NetCDF
+
+
+class Selection(_Engine):
+    """Crop / sel / subset / reduce collaborator for ``NetCDF``."""
+
+    def crop(
+        self,
+        mask: Any = None,
+        touch: bool = True,
+        *,
+        bbox: tuple[float, float, float, float] | list[float] | None = None,
+        epsg: Any = None,
+        chunks: Any = None,
+    ) -> NetCDF:
+        """Crop the dataset using a polygon mask, a raster mask, or a bbox tuple.
+
+        On a **root MDIM container** this crops every variable and
+        returns a new in-memory NetCDF container with the cropped
+        results. On a **variable subset** it delegates to the parent
+        :meth:`pyramids.dataset.Dataset.crop` and re-wraps the result
+        as :class:`NetCDF` to preserve variable metadata
+        (``_band_dim_name``, ``_band_dim_values``, :meth:`sel`).
+
+        Args:
+            mask: GeoDataFrame with polygon geometry, or a Dataset
+                to use as a spatial mask. Mutually exclusive with
+                ``bbox``; exactly one of the two must be supplied.
+            touch: If True, include cells that touch the mask
+                boundary. Defaults to True.
+            bbox (keyword-only): ``(west, south, east, north)``
+                quadruple in the CRS named by ``epsg``. Internally
+                wrapped in a one-row :class:`FeatureCollection` via
+                :meth:`FeatureCollection.from_bbox` and routed through
+                the same polygon path. The FC is built **once** so a
+                root-container crop does not rebuild it for every
+                variable. Mutually exclusive with ``mask``.
+            epsg (keyword-only): CRS for ``bbox`` — anything geopandas
+                accepts for ``crs=`` (EPSG int, ``"EPSG:4326"``, WKT,
+                :class:`pyproj.CRS`). Defaults to the dataset's own
+                CRS, so a bbox in the dataset's native CRS needs no
+                extra argument; pass it explicitly for a bbox in a
+                different CRS (the standard reprojection path handles
+                it).
+            chunks (keyword-only): Lazy-read chunking for the
+                **curvilinear** crop path only — forwarded to
+                :meth:`read_array` so the cropped window is read through
+                the dask-backed lazy path (``"auto"`` or a ``{"rows":
+                ..., "cols": ...}`` dict). The curvilinear crop reads only
+                the polygon's bounding window regardless; ``chunks`` makes
+                that windowed read lazy/chunked. It is a per-variable,
+                curvilinear-only option: it raises ``ValueError`` if given
+                for a rectilinear (affine-warp) crop (which is eager), or
+                on a **root container** (call crop on a single variable via
+                :meth:`get_variable` instead).
+
+        Returns:
+            NetCDF: Cropped container or variable subset.
+
+        Raises:
+            ValueError: Both ``mask`` and ``bbox`` were supplied, or
+                ``chunks`` was given on a root container or a rectilinear
+                crop.
+            TypeError: Neither ``mask`` nor ``bbox`` was supplied.
+
+        Examples:
+            - Crop every variable of a root NetCDF container by a
+              bbox in the dataset's own CRS (`epsg` is inferred). The
+              noah fixture's geotransform is ``cell_size=0.5°``,
+              ``origin=(0, 90)``, 512×512 cells — so its coordinate
+              range is ``x ∈ [0, 256)`` and ``y ∈ (-166, 90]``. The
+              bbox below sits well inside that range:
+                ```python
+                >>> from pyramids.netcdf import NetCDF
+                >>> nc = NetCDF.read_file(
+                ...     "tests/data/netcdf/noah-precipitation-1979.nc"
+                ... )
+                >>> cropped = nc.crop(bbox=(10.0, -50.0, 50.0, -20.0))
+                >>> sorted(cropped.variables) == sorted(nc.variables)
+                True
+
+                ```
+            - Mutual-exclusion guard:
+                ```python
+                >>> from pyramids.feature import FeatureCollection
+                >>> from pyramids.netcdf import NetCDF
+                >>> nc = NetCDF.read_file(
+                ...     "tests/data/netcdf/noah-precipitation-1979.nc"
+                ... )
+                >>> fc = FeatureCollection.from_bbox(
+                ...     (10.0, -50.0, 50.0, -20.0), epsg=nc.epsg,
+                ... )
+                >>> try:
+                ...     nc.crop(mask=fc, bbox=(10.0, -50.0, 50.0, -20.0))
+                ... except ValueError as exc:
+                ...     print("not both" in str(exc))
+                True
+
+                ```
+
+        See Also:
+            - :meth:`pyramids.dataset.Dataset.crop`: same ``bbox=`` /
+              ``epsg=`` surface for plain rasters.
+            - :meth:`pyramids.feature.FeatureCollection.from_bbox`: the
+              shared primitive that builds the one-row FC.
+        """
+        nc = self._ds
+        mask = self._resolve_crop_mask(mask, bbox, epsg)
+        if nc._is_md_array and not nc._is_subset and nc.band_count == 0:
+            # A container crops every variable; `chunks` is a curvilinear-only, per-variable knob
+            # (a container may mix curvilinear and rectilinear variables), so the per-variable fan-out
+            # cannot honour it. Reject it explicitly rather than silently reading eagerly.
+            if chunks is not None:
+                raise ValueError(
+                    "crop(chunks=…) is not supported on a root container; it is a "
+                    "curvilinear-only, per-variable option — call crop on a single variable "
+                    "via get_variable(name) instead."
+                )
+            result = nc._apply_to_all_variables("crop", {"mask": mask, "touch": touch})
+        else:
+            result = self._crop_one(mask, touch=touch, chunks=chunks)
+        return result
+
+    def _resolve_crop_mask(
+        self,
+        mask: Any,
+        bbox: tuple[float, float, float, float] | list[float] | None,
+        epsg: Any,
+    ) -> Any:
+        """Resolve the crop selector to a single polygon mask.
+
+        Converts a ``bbox`` (in ``epsg``, defaulting to the dataset CRS) into a one-row
+        :class:`FeatureCollection`, enforces that exactly one of ``mask`` / ``bbox`` is supplied,
+        and returns the mask to crop with.
+
+        Args:
+            mask: A polygon mask or ``Dataset`` mask, or ``None`` when ``bbox`` is used.
+            bbox: ``(west, south, east, north)`` quadruple, or ``None`` when ``mask`` is used.
+            epsg: CRS for ``bbox``; defaults to ``self.epsg`` when ``None``.
+
+        Returns:
+            The resolved mask (a ``FeatureCollection`` when built from ``bbox``).
+
+        Raises:
+            ValueError: Both ``mask`` and ``bbox`` were supplied, or a ``bbox`` was given with no CRS.
+            TypeError: Neither ``mask`` nor ``bbox`` was supplied.
+        """
+        if bbox is not None:
+            if mask is not None:
+                raise ValueError("crop accepts either `mask` or `bbox`, not both")
+            crs = epsg if epsg is not None else self._ds.epsg
+            if crs is None:
+                raise ValueError(
+                    "crop(bbox=…) requires an explicit `epsg=` when the "
+                    "NetCDF itself has no CRS (self.epsg is None) — a "
+                    "bbox without a CRS is ambiguous"
+                )
+            mask = FeatureCollection.from_bbox(bbox, epsg=crs)
+        if mask is None:
+            raise TypeError(
+                "crop requires a `mask` (GeoDataFrame / FeatureCollection / "
+                "Dataset) or a `bbox` (west, south, east, north) tuple"
+            )
+        return mask
+
+    def _crop_one(self, mask: Any, touch: bool = True, chunks: Any = None) -> "NetCDF":
+        """Crop a single variable/subset, routing curvilinear grids to the 2-D coordinate masker.
+
+        Curvilinear grids (2-D lon/lat coords, no single affine geotransform) can't be clipped by the
+        affine cutline warp, so they mask on their 2-D coordinates; rectilinear grids use the affine
+        crop and re-wrap to preserve NetCDF metadata. ``chunks`` is valid only on the curvilinear path.
+
+        Args:
+            mask: The resolved polygon/raster mask to crop with.
+            touch: If True, include cells touching the mask boundary. Defaults to True.
+            chunks: Lazy-read chunking, valid only on the curvilinear path; see :meth:`crop`.
+
+        Returns:
+            NetCDF: The cropped variable subset.
+
+        Raises:
+            ValueError: ``chunks`` was given for a rectilinear (affine) crop, which is eager.
+        """
+        nc = self._ds
+        curv = NetCDFPlot(nc)._resolve_curvilinear_coords(nc, coords=None)
+        is_curvilinear = (
+            curv is not None
+            and np.asarray(curv[0]).ndim == 2
+            and np.asarray(curv[1]).ndim == 2
+        )
+        if is_curvilinear:
+            result = self._crop_curvilinear(mask, curv, touch=touch, chunks=chunks)
+        else:
+            if chunks is not None:
+                raise ValueError(
+                    "chunks= is only supported for curvilinear crop; the affine "
+                    "(rectilinear) crop path is eager."
+                )
+            # `nc.spatial.crop` is the base Dataset affine crop — exactly what the
+            # NetCDF.crop override reached via `super().crop(...)`, bypassing this engine.
+            result = nc._preserve_netcdf_metadata(
+                nc.spatial.crop(mask=mask, touch=touch)
+            )
+        return result
+
+    def _crop_curvilinear(
+        self,
+        mask: FeatureCollection,
+        coords2d: tuple[np.ndarray, np.ndarray],
+        touch: bool = True,
+        chunks: Any = None,
+    ) -> "NetCDF":
+        """Crop a curvilinear (2-D coordinate) variable by masking on its lon/lat arrays.
+
+        Curvilinear grids have 2-D ``lon(y, x)`` / ``lat(y, x)`` coordinates and no single affine
+        geotransform, so the cutline warp used by :meth:`pyramids.dataset.Dataset.crop` cannot clip
+        them. Instead, test each cell's ``(lon, lat)`` against the polygon, set the cells whose
+        centre falls outside it to no-data, and trim to the bounding ``(row, col)`` index window of
+        the inside cells. The result keeps its windowed 2-D coordinate arrays (stored as
+        ``_curvilinear_coords``) so it stays curvilinear and plots on its real geometry.
+
+        Args:
+            mask (FeatureCollection):
+                Polygon mask (a ``FeatureCollection`` / ``GeoDataFrame``). Its CRS is reconciled
+                with the variable's CRS before the point-in-polygon test.
+            coords2d (tuple[np.ndarray, np.ndarray]):
+                The variable's 2-D ``(lon, lat)`` coordinate arrays, shaped like its spatial dims.
+            touch (bool):
+                Accepted for signature parity with the affine crop. The curvilinear path tests cell
+                centres, so this currently has no effect. Defaults to True.
+
+        Returns:
+            NetCDF: The masked + windowed variable subset, carrying its windowed 2-D coordinates.
+
+        Raises:
+            ValueError: If the polygon does not overlap the grid (no cell centre inside it).
+        """
+        nc = self._ds
+        lon2d = np.asarray(coords2d[0], dtype=float)
+        lat2d = np.asarray(coords2d[1], dtype=float)
+
+        gdf = mask
+        gdf_crs = getattr(gdf, "crs", None)
+        if gdf_crs is not None and nc.epsg:
+            src_epsg = gdf_crs.to_epsg()
+            if src_epsg is not None and src_epsg != nc.epsg:
+                gdf = gdf.to_crs(epsg=nc.epsg)
+        geometry = gdf.geometry.union_all()
+
+        inside = contains_xy(geometry, lon2d, lat2d)
+        if not bool(np.any(inside)):
+            raise ValueError(
+                "crop polygon does not overlap the curvilinear grid "
+                "(no cell centre falls inside it)."
+            )
+
+        rows = np.nonzero(np.any(inside, axis=1))[0]
+        cols = np.nonzero(np.any(inside, axis=0))[0]
+        r0, r1 = int(rows[0]), int(rows[-1]) + 1
+        c0, c1 = int(cols[0]), int(cols[-1]) + 1
+
+        nd = nc.no_data_value
+        nd = nd[0] if isinstance(nd, (list, tuple)) else nd
+        if nd is None:
+            nd = DEFAULT_NO_DATA_VALUE
+
+        # Read only the bounding window, not the whole variable. The polygon mask and window were
+        # derived from the (spatial-footprint) coordinate arrays alone — no data was materialised yet.
+        inside_win = inside[r0:r1, c0:c1]
+        if chunks is not None:
+            # Lazy/chunked read: only the chunks overlapping the window are materialised.
+            lazy = nc.read_array(chunks=chunks)
+            if lazy.ndim > 2:
+                # read_array(chunks=) keeps the native (d0, ..., rows, cols) shape; flatten the
+                # non-spatial dims to the (bands, rows, cols) layout the eager path uses.
+                lazy = lazy.reshape(-1, *lazy.shape[-2:])
+            data_win = np.array(lazy[..., r0:r1, c0:c1].compute(), copy=True)
+        else:
+            # Eager windowed read: GDAL reads just the (c0, r0)-(c1, r1) block.
+            data_win = np.array(
+                nc.read_array(window=[c0, r0, c1 - c0, r1 - r0]), copy=True
+            )
+        data_win[..., ~inside_win] = nd
+        lon_win = lon2d[r0:r1, c0:c1]
+        lat_win = lat2d[r0:r1, c0:c1]
+
+        var_name = getattr(nc, "_source_var_name", None) or "data"
+        container = nc.create_from_array(
+            data_win,
+            geo=nc._bbox_geotransform(lon_win, lat_win),
+            epsg=nc.epsg or 4326,
+            no_data_value=nd,
+            variable_name=var_name,
+        )
+        # create_from_array returns a root container; hand back the variable subset, carrying the
+        # windowed 2-D coordinates so the result stays curvilinear (plots on its real geometry).
+        result = container.get_variable(var_name)
+        result._curvilinear_coords = (lon_win, lat_win)
+        return result
